Resolve agents not terminating even after received stop_msg_confirm
The bug was due to the inner while-loop for round sync.
If an agent finished a round faster and enter round sync for next round.
The `msg_handle` will consume `stop_msg_confirm` but not terminate the thread.
If other agents already stopped, then this agent will stay in the loop forever.

The fix essentially merged the inner loop into the outer loop,
so `agent_gvh.is_alive` is checked for stopping the thread in round sync phases.